chore(release): remove S3 upload and in-game version check
Drop the AWS-backed version.json publish path and the runtime HTTP
update check. Keep version.json for GitHub Release notes and sync CI.
