+++ .github/workflows/release.yml
@@ -7,7 +7,6 @@ on:
 
 permissions:
   contents: write
-  id-token: write
 
 jobs:
   release:
@@ -94,12 +93,3 @@ jobs:
           fail_on_unmatched_files: true
           draft: false
           prerelease: false
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v4
-        with:
-          role-to-assume: ${{ secrets.AWS_ROLE_ARN }}
-          aws-region: eu-west-1
-
-      - name: Upload version.json to S3
-        run: aws s3 cp version.json s3://modding/dwm/version.json
