--- src/main/resources/dwm-version.properties
@@ -1,3 +1,0 @@
-# Expanded at build time from gradle.properties (see processResources in build.gradle).
-minecraft_version=${minecraft_version}
-mod_version=${version}
